Dalsze Ciach

+++ enc_temp_folder/3fc5344a564242aad9f3413fbc9174fb/BaseBuildActor.h
@@ -27,7 +27,7 @@ protected:
 	int32 BuildPoints = 0;
 
 public:
-	virtual void OnPickUp(float bonus=1) PURE_VIRTUAL(ABaseBuildActor::OnPickUp, );
+	virtual void OnPickUp(int32 bonus=1) PURE_VIRTUAL(ABaseBuildActor::OnPickUp, );
 
 	// Called every frame
 	virtual void Tick(float DeltaTime) override;
