+++ enc_temp_folder/a9d2c8ffd687ca9ecf6dd510fb6a333/RedBuildActor.h
@@ -17,5 +17,5 @@ class CITYBUILD_API ARedBuildActor : public ABaseBuildActor
 		
 public:
 	ARedBuildActor();
-	virtual void OnPickUp(float bonus = 1) override;
+	virtual void OnPickUp(int32 bonus = 1) override;
 };
